+++ src/Engine/PostProcessing/PostProcessing.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-class RenderTarget;
 namespace Video {
+    class RenderTarget;
     class Filter;
     namespace Geometry {
         class Rectangle;
@@ -21,7 +21,7 @@ class PostProcessing {
         /**
          * @return Target to render to.
          */
-        RenderTarget* GetRenderTarget() const;
+        Video::RenderTarget* GetRenderTarget() const;
 
         /// Updates the buffer size.
         void UpdateBufferSize();
@@ -41,7 +41,7 @@ class PostProcessing {
     private:
         short int which = 0;
         
-        RenderTarget* buffers[2];
+        Video::RenderTarget* buffers[2];
         
         Video::Geometry::Rectangle* rectangle;
 };
